Category retrievement
Retrieves the categories links and parses them in another method.
For now only write them down in a temp file.

NB: Auchan's Robots.txt doesn't allow some categories to be scraped (Promos at the time of the test for example). We can ignore it by setting the ROBOTSTXT_OBEY variable in settings.py to False but it would violate Auchan ToS (Terms of Service) and we could be open to a lawsuit however unlikely as it is.